Move gameplay logic into its own file
